server/requirements.txt
# python-telegram-bot==20.7

google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

python-dotenv==1.0.0
python-multipart>=0.0.6

requests
Flask==3.1.2
gunicorn>=23.0
redis==6.4.0
